Moving rpc_test utility functions to a separate common.hpp/common.cpp file
