+++ nano/rpc_test/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(rpc_test entry.cpp rpc.cpp)
+add_executable(rpc_test common.hpp common.cpp entry.cpp rpc.cpp)
 
 target_link_libraries(rpc_test node secure rpc test_common gtest)
